# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ResponseData_any_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AttachmentDTO": {
				"properties": {
					"attachmentId": {
						"type": "number",
						"format": "double"
					},
					"messageId": {
						"type": "number",
						"format": "double"
					},
					"fileName": {
						"type": "string"
					},
					"fileUrl": {
						"type": "string"
					},
					"fileType": {
						"type": "string"
					},
					"fileSize": {
						"type": "number",
						"format": "double"
					},
					"mimeType": {
						"type": "string"
					},
					"cloudFileId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"attachmentId",
					"messageId",
					"fileName",
					"fileUrl",
					"fileType",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessageDTO": {
				"properties": {
					"messageId": {
						"type": "number",
						"format": "double"
					},
					"conversationId": {
						"type": "number",
						"format": "double"
					},
					"senderId": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"messageType": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"attachments": {
						"items": {
							"$ref": "#/components/schemas/AttachmentDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"messageId",
					"conversationId",
					"senderId",
					"content",
					"messageType",
					"status",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseData_MessageDTO-Array_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/MessageDTO"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseData_MessageDTO_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/MessageDTO"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendMessageInput": {
				"properties": {
					"conversationId": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"messageType": {
						"type": "string"
					}
				},
				"required": [
					"conversationId",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EditMessageInput": {
				"properties": {
					"content": {
						"type": "string"
					}
				},
				"required": [
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseData__success-boolean__": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"properties": {
							"success": {
								"type": "boolean"
							}
						},
						"required": [
							"success"
						],
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseData__downloadUrl-string--fileName-string--mimeType-string--fileSize-number__": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"properties": {
							"fileSize": {
								"type": "number",
								"format": "double"
							},
							"mimeType": {
								"type": "string"
							},
							"fileName": {
								"type": "string"
							},
							"downloadUrl": {
								"type": "string"
							}
						},
						"required": [
							"fileSize",
							"mimeType",
							"fileName",
							"downloadUrl"
						],
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "tro-tot-vn-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "lamdx4"
		}
	},
	"paths": {
...
		"/chat/conversations/{conversationId}/messages": {
			"get": {
				"operationId": "GetMessages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_MessageDTO-Array_"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Get messages for a conversation",
				"tags": [
					"Chat Messages"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "conversationId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "SendMessage",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_MessageDTO_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Send a new message",
				"tags": [
					"Chat Messages"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "conversationId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendMessageInput"
							}
						}
					}
				}
			}
		},
		"/chat/messages/{messageId}": {
			"get": {
				"operationId": "GetMessageById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_MessageDTO_"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Get a single message by ID",
				"tags": [
					"Chat Messages"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "messageId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "EditMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_MessageDTO_"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Edit a message",
				"tags": [
					"Chat Messages"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "messageId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EditMessageInput"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMessage",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Delete a message",
				"tags": [
					"Chat Messages"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "messageId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chat/messages/read": {
			"post": {
				"operationId": "MarkRead",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData__success-boolean__"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Mark messages as read",
				"tags": [
					"Chat Messages"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"messageIds": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									}
								},
								"required": [
									"messageIds"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/chat/conversations/{conversationId}/files": {
			"post": {
				"operationId": "UploadFile",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_MessageDTO_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"500": {
						"description": "File Upload Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Upload file and send as message",
				"tags": [
					"Chat Messages"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "conversationId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									},
									"content": {
										"type": "string"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/chat/files/{fileId}/download": {
			"get": {
				"operationId": "GetFileDownloadUrl",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData__downloadUrl-string--fileName-string--mimeType-string--fileSize-number__"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Get file download URL",
				"tags": [
					"Chat Messages"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "fileId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/files/{fileId}": {
			"get": {
				"operationId": "GetFile",
				"responses": {
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "File Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Get a file by ID.\nServes the file as a stream with the correct Content-Type.",
				"tags": [
					"Multimedia"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "fileId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}